requirements.txt
asttokens==3.0.0
attrs==19.3.0
backcall==0.2.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.57.0
gudhi==3.11.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
packaging==25.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
python-dateutil==2.9.0.post0
pyzmq==26.4.0
scikit-learn==1.3.2
scipy==1.10.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13\